2023.08.03 commit - application.yaml을 환경별(spring.profiles.active)로 분리 - linkednest regacy backend code 일부 이관 등

+++ src/main/java/net/linkednest/common/service/UserService.java
@@ -1,4 +1,4 @@
-package net.linkednest.www.service;
+package net.linkednest.common.service;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,48 +6,53 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.linkednest.backoffice.dto.menu.ResAdminMenuCategoryDto;
-import net.linkednest.common.dto.user.signin.ReqUserTokenLoginDto;
-import net.linkednest.common.exception.CustomException;
-import net.linkednest.common.utils.ReCaptchaUtil;
+import net.linkednest.common.CommonConstants;
+import net.linkednest.common.ResponseCodeMsg;
 import net.linkednest.common.dto.authority.ResAdminMenuRoleAccessPathDto;
 import net.linkednest.common.dto.authority.ResRoleDto;
 import net.linkednest.common.dto.authority.ResUserRoleAccessPathDto;
 import net.linkednest.common.dto.authority.ResUserRoleDto;
+import net.linkednest.common.dto.user.ResTokenDto;
+import net.linkednest.common.dto.user.get.ResUserDto;
+import net.linkednest.common.dto.user.signin.ReqUserLoginDto;
+import net.linkednest.common.dto.user.signin.ReqUserTokenLoginDto;
+import net.linkednest.common.dto.user.signin.ResUserLoginDto;
+import net.linkednest.common.dto.user.signup.ReqUserRegistDto;
+import net.linkednest.common.dto.user.signup.ResUserProfileDto;
 import net.linkednest.common.entity.content.Content;
 import net.linkednest.common.entity.menu.AdminMenu;
 import net.linkednest.common.entity.menu.AdminMenuCategory;
 import net.linkednest.common.entity.role.AdminMenuRoleAccessPath;
 import net.linkednest.common.entity.role.Authority;
+import net.linkednest.common.entity.role.Role;
+import net.linkednest.common.entity.role.RoleAccessPath;
 import net.linkednest.common.entity.user.LoggedIn;
 import net.linkednest.common.entity.user.User;
+import net.linkednest.common.entity.user.UserProfile;
 import net.linkednest.common.entity.user.UserRefreshToken;
 import net.linkednest.common.repository.ContentRepository;
-import net.linkednest.common.repository.admin.AdminMenuRoleAccessPathRepository;
-import net.linkednest.common.CommonConstants;
-import net.linkednest.common.ResponseCodeMsg;
 import net.linkednest.common.repository.RoleRepository;
+import net.linkednest.common.repository.admin.AdminMenuRoleAccessPathRepository;
 import net.linkednest.common.repository.auth.UserRefreshTokenRepository;
 import net.linkednest.common.repository.user.LoggedInRepository;
+import net.linkednest.common.repository.user.UserProfileRepository;
 import net.linkednest.common.repository.user.UserRepository;
-import net.linkednest.common.dto.user.ResTokenDto;
-import net.linkednest.common.dto.user.signin.ReqUserLoginDto;
-import net.linkednest.common.dto.user.signin.ResUserLoginDto;
-import net.linkednest.common.dto.user.signup.ReqUserRegistDto;
 import net.linkednest.common.security.CustomUserDetails;
 import net.linkednest.common.security.JwtProvider;
 import net.linkednest.common.utils.CommonUtil;
+import net.linkednest.common.utils.ReCaptchaUtil;
+import net.linkednest.share.service.ShareService;
+import net.linkednest.www.service.UserProfileService;
 import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Sort;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import java.io.IOException;
 import java.util.*;
 
 @Slf4j
@@ -55,7 +60,9 @@ import java.util.*;
 @RequiredArgsConstructor
 public class UserService {
     private final UserProfileService                    userProfileService;
+    private final ShareService                          shareService;
     private final UserRepository                        userRepository;
+    private final UserProfileRepository                 userProfileRepository;
     private final RoleRepository                        roleRepository;
     private final UserRefreshTokenRepository            userRefreshTokenRepository;
     private final AdminMenuRoleAccessPathRepository     adminMenuRoleAccessPathRepository;
@@ -66,14 +73,14 @@ public class UserService {
 
     @Transactional(rollbackOn = {DataIntegrityViolationException.class})
     public Boolean registUser(ReqUserRegistDto userRegistDto) {
-        String methodName = new Object(){}.getClass().getEnclosingMethod().getName();
-        String reCaptchaToken = StringUtils.defaultString(userRegistDto.getReCaptchaToken());
+        String  methodName          = new Object(){}.getClass().getEnclosingMethod().getName();
+        String  reCaptchaToken      = StringUtils.defaultString(userRegistDto.getReCaptchaToken());
         boolean isReCaptchaVerified = ReCaptchaUtil.verifyV3(reCaptchaToken);
 
         if (isReCaptchaVerified) {
             log.info("[{}.{}] userRegistObj : {}", this.getClass().getName(), methodName, userRegistDto.toString());
-            User newUser = new User();
-
+            User newUser        = new User();
+            User createdUser    = null;
             try {
                 newUser.setUserId(new String(Base64.getDecoder().decode(userRegistDto.getUserId())));
                 newUser.setEmail(userRegistDto.getEmail());
@@ -87,23 +94,15 @@ public class UserService {
                 newUser.setCreateDate(new Date());
 
                 log.info("[{}.{}] userRegist >> decoded : {}", this.getClass().getName(), methodName, newUser.toString());
+                createdUser = userRepository.saveAndFlush(newUser);
+                log.info("[{}.{}] createdUser : {}", this.getClass().getName(), methodName, createdUser);
 
-                try {
-                    User createdUser = userRepository.saveAndFlush(newUser);
-                    log.info("[{}.{}] createdUser : {}", this.getClass().getName(), methodName, createdUser);
-                    try {
-                        this.userProfileService.saveUserProfile(userRegistDto, createdUser);
-                    } catch (Exception e) {
-                        userRepository.delete(createdUser);
-                        log.error("[{}.{}] userProfileService.saveUserProfile ERROR OCCURRED reqDto : {}, errorMsg : {}", this.getClass().getName(), methodName, userRegistDto.toString(), e.getMessage());
-                        return false;
-                    }
-                } catch (Exception e) {
-                    log.error("[{}.{}] userRepository.save ERROR OCCURRED entity : {}, errorMsg : {}", this.getClass().getName(), methodName, newUser, e.getMessage());
-                    return false;
-                }
+                this.userProfileService.saveUserProfile(userRegistDto, createdUser);
+                this.shareService.mergeShare(createdUser);
             } catch (Exception e) {
-                log.error("[{}.{}] userRepository.save ERROR OCCURRED entity : {}, errorMsg : {}", this.getClass().getName(), methodName, newUser, e.getMessage());e.printStackTrace();
+                if (createdUser != null) userRepository.delete(createdUser);
+                log.error("[{}.{}] ERROR OCCURRED entity : {}, errorMsg : {}", this.getClass().getName(), methodName, newUser, e.getMessage());
+                e.printStackTrace();
                 return false;
             }
         } else {
@@ -114,9 +113,9 @@ public class UserService {
 
     public Boolean updateUser(ReqUserRegistDto userRegistDto) {
         log.info("[{}.{}] updateUser : {}", this.getClass().getName(), "updateUser", userRegistDto.toString());
-        String  decodedUserId   = new String(Base64.getDecoder().decode(userRegistDto.getUserId()));
+        String  decodedUserId = new String(Base64.getDecoder().decode(userRegistDto.getUserId()));
         try {
-            Optional<User> userOptional = this.getUser(decodedUserId);
+            Optional<User> userOptional = this.findByUserId(decodedUserId);
             if (userOptional.isPresent()) {
                 User user = userOptional.get();
                 user.setUserId(decodedUserId);
@@ -148,7 +147,7 @@ public class UserService {
             String userId   = new String(Base64.getDecoder().decode(reqUserLoginDto.getUserId()));
             String password = new String(Base64.getDecoder().decode(reqUserLoginDto.getPassword()));
 
-            Optional<User> userOptional = this.getUser(userId);
+            Optional<User> userOptional = this.findByUserId(userId);
             if (userOptional.isPresent()) {
                 user = userOptional.get();
                 // password check
@@ -189,7 +188,6 @@ public class UserService {
         } else {
             returnCode = 20003;
         }
-
         resUserLoginDto.setReturnCode(returnCode);
         resUserLoginDto.setReturnMsg(ResponseCodeMsg.of(returnCode).getResMsg());
         resUserLoginDto.setIsLogin(returnCode == 10000);
@@ -197,7 +195,6 @@ public class UserService {
         if (ObjectUtils.isNotEmpty(user) && returnCode == 10000) {
             this.setLoggedIn(request, user, CommonConstants.WEB_LOGIN_ACTION);
         }
-
         return resUserLoginDto;
     }
 
@@ -218,15 +215,14 @@ public class UserService {
                 // accessToken 확인
                 accessToken = accessToken.split(" ")[1].trim();
                 log.info("[{}.{}] accessToken 확인 : {}", this.getClass().getName(), "tokenLogin", accessToken);
-                String tokenUserId  = jwtProvider.getUserId(accessToken);
-
+                String tokenUserId = jwtProvider.getUserId(accessToken);
                 log.info("[{}.{}] tokenUserId : {}, paramUserId : {}", this.getClass().getName(), "tokenLogin", tokenUserId, userId);
 
-                Optional<User> userOptional = this.getUser(userId);
+                Optional<User> userOptional = this.findByUserId(userId);
                 if (userOptional.isPresent()) {
                     user = userOptional.get();
                     // accessToken 체크
-                    boolean isValidUserId      = StringUtils.equals(userId, tokenUserId);
+                    boolean isValidUserId = StringUtils.equals(userId, tokenUserId);
                     if (isValidToken && isValidUserId) {
                         resUserLoginDto.setIsLogin(true);
                         resUserLoginDto.setUserNo(user.getUserNo());
@@ -252,7 +248,6 @@ public class UserService {
 
                         // refresh token 발급(or 재사용)
                         UserRefreshToken mergedRefreshToken = this.mergeRefreshToken(user);
-
                         response.setHeader(CommonConstants.REFRESH_TOKEN, mergedRefreshToken.getRefreshToken());
                     } else {
                         returnCode = 20001;
@@ -289,9 +284,9 @@ public class UserService {
 
     @Async
     public void setLoggedIn(HttpServletRequest request, User user, String actionType) {
-        String ipAddr = CommonUtil.getClientIp(request);
-        String refererUrl = request.getHeader("referer");
-        LoggedIn loggedIn = new LoggedIn();
+        String      ipAddr      = CommonUtil.getClientIp(request);
+        String      refererUrl  = request.getHeader("referer");
+        LoggedIn    loggedIn    = new LoggedIn();
 
         if (StringUtils.isNotEmpty(refererUrl)) {
             String[] pathArr = refererUrl.split("/");
@@ -341,24 +336,22 @@ public class UserService {
         String methodName = new Object(){}.getClass().getEnclosingMethod().getName();
         amrapList.forEach(a -> {
             log.info("[{}.{}] AdminMenuRoleAccessPath : {}", this.getClass().getName(), methodName, a.getAdminMenu().getMenuName());
-            AdminMenu am  = a.getAdminMenu();
-            AdminMenuCategory amc = am.getAdminMenuCategory();
-            String              categoryName = amc.getCategoryName();
+            AdminMenu         am           = a.getAdminMenu();
+            AdminMenuCategory amc          = am.getAdminMenuCategory();
+            String            categoryName = amc.getCategoryName();
             List<ResAdminMenuRoleAccessPathDto> resAdminMenuRoleAccessPathDtoList   = new ArrayList<>();
             Optional<ResAdminMenuCategoryDto>   resAdminMenuCategoryDtoOptional     = adminMenuCategoryDtoList.stream().filter(amcdl -> amcdl.getCategoryId().equals(amc.getId())).findAny();
             ResAdminMenuCategoryDto             resAdminMenuCategoryDto             = null;
 
             log.info("[{}.{}] resAdminMenuCategoryDtoOptional : {}", this.getClass().getName(), "setAdminMenuCategoryList", resAdminMenuCategoryDtoOptional.isPresent());
-
             boolean isAddedCategory = resAdminMenuCategoryDtoOptional.isPresent();
 
             if (isAddedCategory) {
-                resAdminMenuCategoryDto = resAdminMenuCategoryDtoOptional.get();
-                resAdminMenuRoleAccessPathDtoList = resAdminMenuCategoryDtoOptional.get().getAdminMenuRoleAccessPathList();
+                resAdminMenuCategoryDto             = resAdminMenuCategoryDtoOptional.get();
+                resAdminMenuRoleAccessPathDtoList   = resAdminMenuCategoryDtoOptional.get().getAdminMenuRoleAccessPathList();
                 adminMenuCategoryDtoList.remove(resAdminMenuCategoryDto);
 
                 log.info("[{}.{}] resAdminMenuRoleAccessPathDtoList : {}", this.getClass().getName(), "setAdminMenuCategoryList", resAdminMenuRoleAccessPathDtoList);
-
             } else {
                 resAdminMenuCategoryDto = new ResAdminMenuCategoryDto();
                 resAdminMenuCategoryDto.setCategoryId(amc.getId());
@@ -408,7 +401,7 @@ public class UserService {
         return mergedRefreshToken;
     }
 
-    public Optional<User> getUser(String userId) {
+    public Optional<User> findByUserId(String userId) {
         return userRepository.findByUserId(userId);
     }
 
@@ -426,5 +419,70 @@ public class UserService {
         resTokenDto.setReturnCode(userRefreshTokenOptional.isPresent() ? 10000 : 50000);
         resTokenDto.setReturnMsg(userRefreshTokenOptional.isPresent() ? ResponseCodeMsg.of(10000).getResMsg() : "FAIL");
         return resTokenDto;
+    }
+
+    public List<ResUserDto> userList() {
+        List<ResUserDto> userList = new ArrayList<>();
+        userRepository.findAll().forEach(u -> {
+            userList.add(this.getUser(u));
+        });
+        return userList;
+    }
+    public ResUserDto getUser(String userId) {
+        Optional<User> userOptional = userRepository.findByUserId(userId);
+        return userOptional.map(this::getUser).orElse(null);
+    }
+    private ResUserDto getUser(User u) {
+        ResUserDto userObj = new ResUserDto();
+        userObj.setUserNo(u.getUserNo());
+        userObj.setUserId(u.getUserId());
+        userObj.setEmail(u.getEmail());
+        userObj.setNickname(u.getNickname());
+        userObj.setIntroduce((u.getIntroduce()));
+        userObj.setUserProfile(this.getUserProfile(u));
+
+        List<ResUserRoleDto> userRoleList = new ArrayList<>();
+        u.getUserRoles().forEach(r -> {
+            userRoleList.add(this.getUserRole(r));
+        });
+        userObj.setUserRoleInfoList(userRoleList);
+
+        return userObj;
+    }
+    private ResUserProfileDto getUserProfile(User u) {
+        Optional<UserProfile> upOptional = userProfileRepository.findByUser(u);
+        ResUserProfileDto userProfileObj = new ResUserProfileDto();
+        if (upOptional.isPresent()){
+            UserProfile up = upOptional.get();
+            userProfileObj.setSex(up.getSex());
+            userProfileObj.setBirthday(up.getBirthday());
+            userProfileObj.setPhoneNo(up.getPhoneNo());
+        }
+        return userProfileObj;
+    }
+    private ResUserRoleDto getUserRole(Authority r) {
+        ResUserRoleDto userRoleObj = new ResUserRoleDto();
+        userRoleObj.setRoleName(r.getRole().getRoleName());
+        userRoleObj.setRoleId(r.getRole().getId());
+        // user menu role(for back-end)
+        Role role = r.getRole();
+        List<ResUserRoleAccessPathDto> userRoleAccessPathList = new ArrayList<>();
+        role.getAccessPaths().forEach(ap -> {
+            userRoleAccessPathList.add(this.getUserRoleAccessPath(ap));
+        });
+        userRoleObj.setUserRoleAccessPathList(userRoleAccessPathList);
+
+        // user menu role(for front-end)
+        List<ResAdminMenuCategoryDto> adminMenuCategoryList = new ArrayList<>();
+        this.setAdminMenuCategoryList(r, adminMenuCategoryList);
+        userRoleObj.setAdminMenuCategoryList(adminMenuCategoryList);
+        return userRoleObj;
+    }
+    private ResUserRoleAccessPathDto getUserRoleAccessPath(RoleAccessPath ap) {
+        ResUserRoleAccessPathDto userRoleAccessPathObj = new ResUserRoleAccessPathDto();
+        userRoleAccessPathObj.setHttpMethod(ap.getHttpMethod());
+        userRoleAccessPathObj.setUrl(ap.getUrl());
+        userRoleAccessPathObj.setType(ap.getType());
+        return userRoleAccessPathObj;
     }
 }
